# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Online Examination Proctoring System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enrollments/{id}/violations:
    get:
      operationId: listViolations
      tags: [violations]
      summary: Get violations for enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
    post:
      operationId: reportViolation
      tags: [violations]
      summary: Report a proctoring violation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViolationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
...
components:
  schemas:
...
    Violation:
      type: object
      required: [id, enrollmentId, type, severity, timestamp]
      properties:
        id:
          type: integer
        enrollmentId:
          type: integer
        type:
          type: string
          enum: [face_missing, multiple_faces, eye_deviation, audio_detected, tab_switch, fullscreen_exit, copy_paste, phone_detected]
        severity:
          type: string
          enum: [low, medium, high, critical]
        details:
          type: ["string", "null"]
        timestamp:
          type: string
    ViolationInput:
      type: object
      required: [type, severity]
      properties:
        type:
          type: string
          enum: [face_missing, multiple_faces, eye_deviation, audio_detected, tab_switch, fullscreen_exit, copy_paste, phone_detected]
        severity:
          type: string
          enum: [low, medium, high, critical]
        details:
          type: string
...