# lib/api-zod/src/generated/types/violationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Online Examination Proctoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { ViolationInputSeverity } from './violationInputSeverity';
import type { ViolationInputType } from './violationInputType';

export interface ViolationInput {
  type: ViolationInputType;
  severity: ViolationInputSeverity;
  details?: string;
}